refactor: 共通コンポーネントの準備を作成
