--- frontend/src/components/auth/hooks/useSignUp
@@ -1,3 +1,0 @@
-export const useSignUp = () => {
-
-}
